sso_back/requirements1.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
anyio==3.3.4
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.1
asgiref==3.2.10
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2018.1.18
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dictor==0.1.9
diff-match-patch==20200713
distlib==0.3.2
Django==2.2.27
django-ckeditor==6.1.0
django-cors-headers==2.1.0
django-crispy-forms==1.14.0
django-filter==2.4.0
django-formtools==2.1
django-import-export==1.0.0
django-js-asset==1.2.2
django-mixins==0.0.10
django-redis==5.2.0
django-rest-swagger==2.1.2
django-reversion==2.0.13
django-stdimage==5.3.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.0.12
future==0.15.2
generics==3.4.0
h11==0.12.0
httplib2==0.19.1
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.9.1
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonfield==3.1.0
jsonpath==0.82
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
loguru==0.5.3
lxml==4.6.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
outcome==1.1.0
packaging==21.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.11.15
pipreqs==0.4.11
platformdirs==2.3.0
pluggy==0.13.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
py==1.10.0
pycparser==2.20
Pygments==2.11.2
PyMySQL==0.8.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypinyin==0.36.0
pyrsistent==0.18.1
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-repeat==0.9.1
python-crontab==2.2.8
python-dateutil==2.8.2
python-docx==0.8.6
pytz==2017.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
redis==4.1.4
requests==2.26.0
selenium==4.0.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.10.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
sqlparser==0.0.9
starlette==0.17.1
sweet==0.1.12
sweet.web==0.0.9
tablib==3.0.0
terminado==0.13.3
testpath==0.6.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.21.1
uvicorn==0.15.0
virtualenv==20.7.2
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
win32-setctime==1.0.3
wrapt==1.13.3
wsproto==1.0.0
xlrd==2.0.1
XlsxWriter==1.0.5
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
yarg==0.1.9
zipp==3.5.0
